apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "job-hunter FastAPI backend — PDF parsing, LLM extraction, profile storage"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "pymupdf>=1.24.0",
    "pdfminer.six>=20221105",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.12",
    "pydantic-settings>=2.13.1",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
